refactor & add strict schema

+++ Makefile
@@ -1,26 +1,38 @@
 # Set the name of the Docker image used to build the connector
 IMAGE_NAME=kafka-connect-ytsaurus-builder
 
+# Check if JDK is installed and its version is greater than or equal to 11
+JDK_VERSION_MAJOR=$(shell javac -version 2>&1 | awk '{print $$2}' | awk -F '.' '{print $$1}')
+
+ifeq ($(origin USE_GRADLE), undefined)
+  ifeq ($(shell [ -z "$(JDK_VERSION_MAJOR)" ] || [ "$(JDK_VERSION_MAJOR)" -lt "11" ] && echo false || echo true),true)
+    USE_GRADLE=true
+  else
+    USE_GRADLE=false
+  endif
+endif
+
 # Define the build target
 .PHONY: build build-image clean
+
+# Target to build the connector JAR file
+build: build/libs/kafka-connect-ytsaurus.jar
 
 # Target to build the Docker image
 build-image:
 	# Build the Docker image using the specified name
 	docker build -t $(IMAGE_NAME) .
 
-# Target to build the connector JAR file
-build: build/libs/kafka-connect-ytsaurus.jar
-
 # Recipe to build the connector JAR file
-build/libs/kafka-connect-ytsaurus.jar: build-image
+build/libs/kafka-connect-ytsaurus.jar:
+ifeq ($(USE_GRADLE), true)
+	./gradlew shadowJar
+else
+	$(MAKE) build-image
 	# Create a temporary directory to hold the exported Docker image
 	mkdir -p ./tmp
-	# Create a Docker container using the builder image
 	CONTAINER_ID=$$(docker create $(IMAGE_NAME)); \
-	# Export the contents of the container to a tar file
 	docker export $$CONTAINER_ID -o ./tmp/image.tar; \
-	# Remove the container
 	docker rm $$CONTAINER_ID;
 	# Extract the connector JAR file from the exported tar file
 	tar -xf ./tmp/image.tar -C ./tmp app/build/libs/*
@@ -30,6 +42,7 @@ build/libs/kafka-connect-ytsaurus.jar: build-image
 	mv ./tmp/app/build/libs/* build/libs/
 	# Remove the temporary directory
 	rm -rf tmp
+endif
 
 # Target to clean up build artifacts
 clean:
